requirements.txt
streamlit
numpy 
pandas 
matplotlib
transformers
datasets
torch
torchvision
pytesseract
opencv-python
pillow
tqdm
scikit-learn
fuzzywuzzy
python-Levenshtein
ipykernel
sentence-transformers
faiss-cpu
groq